MkImage: Allow pure expressions on right-hand side of enum definitions.
Right-hand side of an enumeration can now be an expression, provided
that it references only pure symbols and literals. There is a residual
bug here, which is that non-integer values can presently appear on
RHS.

This update has been applied in Interp.cxx, but the corresponding
change has not yet been made to ConstDef.cxx. The present change does
not break anything, but it is not yet complete.