capidl: emit server header with dependent headers
Added a new target language -l c-server-all-header that emits headers
for the IDL files that appear on the command line and also the headers
for all of the IDL files on which the command line headers depend.

If I got this right, this *should* yield correct but minimal
server-side header generation.