More progress toward APIC support.
At this point, the basic logic for chip handling is in. If booted with
the "apic" command line argument, we will probe the world and tell you
what is there (this is debugging code, which will shortly be
retired). Having gotten that far, it emerged that we need a
software-managed interrupt source model as well as an interrupt vector
model. I'm about to take a dinner break, and then I will put that in
and we can find out if any of this code really works.