Register interactions with the IOAPIC vector registers now works.
I'm not sure that I am sending the right values, but the interactions
work. With a bit more effort, I *should* now be able to get the
vectors programmed correctly. For the legacy interrupts, I need to
understand whether they are active high or active low, and whether
they should still be configured for edge trigger when operating
through the APIC (I suspect yes).

There is a secondary issue here: should we be using a periodic base
interrupt or should we be setting up the next interval each time the
current interval expires? Going to programmed intervals involves some
logic changes.