Initialize maps properly.
The various mapping tables set up in ldscript.S did not fall within
any allocated region of the ELF file, with the consequence that they
occupied space in the map but were neither initialized nor
zeroed. When running under QEMU and Bochs this caused no problem,
because they were getting loaded into memory that chanced to be
zero.

But when loaded on some real hardware, unknown bits might be set in
the page directory entries with unpredictable results. I suspect this
was the cause of the strangeness we saw in the video modes at
work. Fortunately it manifested via a caught assertion at home: USER
bit set in a page directory entry where it should not have been.

This bug only impacted entries in the page directory table, because
the leaf page table was getting full initialized and the transient map
was getting cleared later (which is now redundant, but I haven't
removed it).

I have not yet tested booting PAE mode on real hardware with this fix.