Re-enable PAE mode
I have checked this on three real machines. Two worked fine. The third
had an unrelated problem in physmem, most likely because it is
provisioned with 4G of physical memory (which leads to some
physical memory provisioning challenges).

I have also checked this under QEMU with the *new* HardClock via lapic
stuff enabled. That also works fine, with no sign of the 0xf reserved
fault. No way to know now, but it seems plausible that we either were
getting the NX bit set or some other reserved bit set in an upper PAE
word through sheer misfortune. What is not obvious (to me) is why that
mattered, since we *were* initializing all of the entries in the PAE
page directory that we were actually loading.

I still think that this was all very weird, but it seems to be fixed now.