Strengthen the ACPI probe logic, permit IRQ overrides.
Kernel now prints all of the ACPI tables found in the RSDT, and
reports bad checksums if appropriate. We actually have a machine that
has an SSDT with a bad checksum, so bad checksums are not fatal unless
they impact the APIC table, which we actually need.

Kernel now processes IRQ overrides from the APIC table correctly.