Add cache control support for physical page capabilities.
I belatedly realized that I needed this for the TextConsole
implementation.

This change adds new permission bits CAP_PERM_CD and CAP_PERM_WT
meaning (respectively "cache disable" and "write through". These bits
can only be set on physical page capabilities. They occupy the same
bit positions as CAP_PERM_OP and CAP_PERM_NC, which are not meaningful
on page capabilities.

A new function nocache() was added in builtins.cxx. While I was there
I added nocall(), which was missing. I added writethrough(), but that
is presently ifdef'd out.

The main change here was in PageFault.c, which sets the PCD and PWT
bits according to the new permissions. The catch is that these
permissions mean something quite different in the leaf MemWalkEntry,
so it is necessary to handle them specially. There was already special
processing going on to make sure the leaf page got dirtied when
required. Hopefully, I put this in the right place, but I need JWA to
review this change.

The rest was merely a matter of updating the PTE installer functions.

I checked the legacy manuals. To my surprise, PCD and PWT were
implemented on the i486, which is the earliest processor we
support. In consequence we don't need to do any legacy processor
checks for these.