Update the PhysMem allocator to use spinlocks.
This subsystem will soon be used to introduce new physical memory
regions to the kernel, so it needs to use protective locking. Adding
that required in turn that the CPU0 structure be constructed earlier
so that it would have a sensible lockword value.

While I was about it, I removed the TY_CPU lock type.