Partial integration of Jeroen's SMP AP startup code.
I started this with the intent of doing a partial integration of
Jeroen Visser's AP startup code, and got sucked in to re-arranging it
a bit. Credit for this work should properly go to Jeroen; any errors
introduced during my mangling are mine alone.

Note that this code does not yet actually start the APs. The IPI
support code is still missing. Hopefully Jeroen can now integrate that
with minimal effort.