Update transmap to handle SMP use-cases.
This is a hand-integration of Jeroen Visser's patch. The substantive
change is a slightly cleaner approach to the ldscript
configuration. The rest is minor style alterations and moving a type
definition from transmap.h to kerntypes.h to avoid adding a new
file. Jeroen's work on this was well thought-out, and has saved all of
us a good bit of time here.

Nice work, Jeroen!