When a method is client-only, omit the IDL_env wrapper anyway.
Client-only methods are actually library functions supplied on the
client side. There is no reason for the library to provide the trivial
wrapper function as well, and many reasons to want to inline
it. Beginning with this commit, the wrapper function is generated even
if the stub is client only.