Quick attempt to fix the HTML back end.
This change splits HTML emission and XML emission, and introduces an
attempt to properly strip the doxygen: and capdoc: tags when going to
HTML. The HTML emitter is untested, but with a bit of luck this will
help Charlie Landau to emit HTML pages from IDL. What it will not
accomplish is cross-linking of identifier references.

That cannot be accomplished without more effort than I am presently
able to spend on it.